Title: 최댓값 구하기, Time: 112ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240301/최댓값 구하기/maximum-value.py
@@ -1,4 +1,4 @@
-a, b, c = map(int, input().split())
+a, b, c = map(float, input().split())
 if a > b:
     if a >= c:
         print(a)
